SA-897 - rename HS abbreviation to Health Score
