+++ src/pages/signals/components/HealthScoreLegend.module.scss
