+++ src/pages/signals/components/HealthScoreMetric.module.scss
