+++ src/pages/signals/components/tests/HealthScoreLegend.test.js
@@ -1,9 +1,9 @@
 import { shallow } from 'enzyme';
 import React from 'react';
-import HSLegend from '../HSLegend';
+import HealthScoreLegend from '../HealthScoreLegend';
 
 describe('Signals Health Score legend Component', () => {
   it('renders correctly', () => {
-    expect(shallow(<HSLegend />)).toMatchSnapshot();
+    expect(shallow(<HealthScoreLegend />)).toMatchSnapshot();
   });
 });
